Some shader stuff

+++ NewEngine/D3D.cpp
@@ -64,7 +64,7 @@ D3D::D3D(int screenWidth, int screenHeight, bool vsync, HWND hwnd, bool fullscre
 	videoCardMemory = (int)(adapterDesc.DedicatedVideoMemory / 1024 / 1024);
 
 	// Convert the name of the video card to a character array and store it.
-	unsigned int stringLength;
+	size_t stringLength;
 	auto error = wcstombs_s(&stringLength, videoCardDescription, 128, adapterDesc.Description, 128);
 	if (error != 0)
 		throw runtime_error("Video Card Name Error");
